requirements.txt
mariadb==1.0.7
pycryptodome==3.10.1